chore: refactoring nodemailer
