+++ helpers/index.js
@@ -1,11 +1,11 @@
-const { sendActivationMail, sendConfirmationMail } = require('./mailer')
+// const { sendActivationMail, sendConfirmationMail } = require('./mailer')
 const { encrypt, confirmToken } = require('./hash')
 
 module.exports = {
-  mailer: { 
-    sendActivationMail,
-    sendConfirmationMail,
-  },
+  // mailer: { 
+  //   sendActivationMail,
+  //   sendConfirmationMail,
+  // },
   hashing: {
     encrypt,
     confirmToken,
