+++ helpers/mailer.js
@@ -1,0 +1,10 @@
+// const { sendActivationMail } = require('./templates/activation')
+// const { sendConfirmationMail } = require('./templates/confirmation')
+
+
+
+
+// module.exports = {
+//   sendActivationMail,
+//   sendConfirmationMail,
+// }
