--- helpers/mailer/index.js
@@ -1,7 +1,0 @@
-const { sendActivationMail } = require('./activation')
-const { sendConfirmationMail } = require('./confirmation')
-
-module.exports = {
-  sendActivationMail,
-  sendConfirmationMail,
-}
